requirements.txt
anytree
cassio
coloredlogs
dash
dash_bootstrap_components
dash-loading-spinners
gliner
html2text
keybert
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
python-dotenv
ragstack-ai-knowledge-store
tabulate
unstructured